backend/requirements.txt
Django==4.2.*
djangorestframework==3.14.*
djangorestframework-simplejwt==5.3.*
django-cors-headers==4.*
mysqlclient==2.2.*
Pillow==10.*
django-environ==0.11.*
cryptography==42.*
pyotp==2.9.*
qrcode[pil]==7.4.*
PyJWT==2.8.*
social-auth-app-django==5.4.*
social-auth-core==4.5.*
pytest-django==4.7.*
pytest-cov==4.*
black==24.*
flake8==7.*
factory-boy==3.3.*
channels==4.0.*
channels-redis==4.2.*
daphne==4.0.*
redis==5.*
openpyxl==3.1.*
tablib==3.5.*
reportlab==4.1.*
pypdf==4.1.*
python-docx==1.1.*
django-dbbackup==4.1.*
celery>=5.3,<6
django-celery-beat>=2.5,<3
